requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
playsound>=1.3.0
elevenlabs>=0.2.27
